Remove custom repo option, now only test/operational from CLI and no-gitlab only as a flag

+++ bipy_gui_manager/create_project/validation.py
@@ -1,9 +1,7 @@
-from typing import Any, Optional, Mapping, Union, Tuple
+from typing import Any, Optional, Tuple
 import os
-import re
 import shutil
 from bipy_gui_manager import cli_utils as cli
-from bipy_gui_manager.create_project import GROUP_NAME
 from bipy_gui_manager.phonebook.phonebook import Phonebook, PhonebookEntry
 
 
@@ -28,50 +26,52 @@ def validate_or_ask(validator, question, neg_feedback, start_value=None, pos_fee
     return value
 
 
-def validate_as_arg_or_ask(cli_value, validator, question, neg_feedback, pos_feedback=None, hints=(), interactive=True)\
-        -> bool:
-    """
-        If an initial value is given and valid, return it.
-        If an initial value is given and invalid, fail.
-        If it's not given, ask the user until a valid value is received.
-        if interactive is False, fail rather than asking.
-    """
-    if not interactive and cli_value is None:
-        raise ValueError(neg_feedback)
-    if cli_value is not None or not interactive:
-        result = validate_or_fail(cli_value, validator, neg_feedback)
-        if pos_feedback:
-            cli.positive_feedback(pos_feedback.format(result))
-        return result
-    else:
-        return validate_or_ask(validator, question, neg_feedback, pos_feedback=pos_feedback, hints=hints)
+# def validate_as_arg_or_ask(cli_value, validator, question, neg_feedback, pos_feedback=None, hints=(), interactive=True)\
+#         -> bool:
+#     """
+#         If an initial value is given and valid, return it.
+#         If an initial value is given and invalid, fail.
+#         If it's not given, ask the user until a valid value is received.
+#         if interactive is False, fail rather than asking.
+#     """
+#     if not interactive and cli_value is None:
+#         raise ValueError(neg_feedback)
+#     if cli_value is not None or not interactive:
+#         result = validate_or_fail(cli_value, validator, neg_feedback)
+#         if pos_feedback:
+#             cli.positive_feedback(pos_feedback.format(result))
+#         return result
+#     else:
+#         return validate_or_ask(validator, question, neg_feedback, pos_feedback=pos_feedback, hints=hints)
 
 
-def resolve_as_arg_or_ask(cli_value, resolver, question, neg_feedback, pos_feedback=None, hints=(), interactive=True)\
-        -> Any:
+def resolve_as_arg_or_ask(initial_value, resolver, question, neg_feedback, pos_feedback=None, hints=(),
+                          interactive=True) -> Any:
     """
         If an initial value is given and valid, resolve and return it.
         If an initial value is given and invalid, fail.
         If it's not given, ask the user until a valid value is received and return it's resolved version.
         if interactive is False, fail rather than asking.
     """
-    if not interactive and cli_value is None:
-        raise ValueError(neg_feedback)
-    if cli_value is not None or not interactive:
-        # Resolve or fail - never ask
-        value = resolver(cli_value)
-        if value is None:
+    # Can't ask for a new value if what's given is wrong
+    if not interactive:
+
+        # Either it's valid or I fail
+        value, success = resolver(initial_value)
+        if not success:
             raise ValueError(neg_feedback)
         if pos_feedback:
             cli.positive_feedback(pos_feedback.format(value))
         return value
+
     else:
-        # Ask and resolve until valid
-        value = None
-        while value is None:
+        # Either it's valid or I ask
+        value, success = resolver(initial_value)
+
+        while not success:
             value = cli.ask_input(question)
-            value = resolver(value)
-            if value is None:
+            value, success = resolver(value)
+            if not success:
                 cli.negative_feedback(neg_feedback)
                 for hint in hints:
                     cli.give_hint(hint)
@@ -80,8 +80,11 @@ def resolve_as_arg_or_ask(cli_value, resolver, question, neg_feedback, pos_feedb
                     cli.positive_feedback(pos_feedback.format(value))
                 return value
 
+        return value
 
-def validate_base_path(base_path: str, project_name: str, interactive: bool = True, overwrite: bool = False) -> bool:
+
+def validate_base_path(base_path: str, project_name: str, interactive: bool = True, overwrite: bool = False) \
+        -> Tuple[Optional[str], bool]:
     """
     Makes sure there is no folder with the name of the new project and if so ask the user and act accordingly.
     :param base_path: path to the new project
@@ -90,12 +93,15 @@ def validate_base_path(base_path: str, project_name: str, interactive: bool = Tr
     :param overwrite: whether to automatically overwrite the existing folder
     :return True if the path is available
     """
+    if base_path is None:
+        return None, False
+
     project_path = os.path.join(base_path, project_name)
     if os.path.exists(project_path):
         if overwrite:
             cli.list_subtask("Overwriting folder {}".format(project_path))
             shutil.rmtree(project_path)
-            return True
+            return base_path, True
         elif not interactive:
             raise OSError("Directory '{}' already exists.".format(project_path))
         else:
@@ -104,9 +110,9 @@ def validate_base_path(base_path: str, project_name: str, interactive: bool = Tr
             if answer == "overwrite":
                 cli.list_subtask("Overwriting existing folder")
                 shutil.rmtree(project_path)
-                return True
-            return False
-    return True
+                return base_path, True
+            return None, False
+    return base_path, True
 
 
 def validate_demo_flags(force_demo: bool, demo: bool, interactive: bool) -> bool:
@@ -145,106 +151,38 @@ def validate_demo_flags(force_demo: bool, demo: bool, interactive: bool) -> bool
                 value = cli.handle_failure("Please type yes or no:")
 
 
-def resolve_gitlab_repo_url(repo_url, upload_protocol, project_name) -> Tuple[Optional[str], bool]:
-
-    if repo_url is None:
-        # Invalid
-        return None, False
-
-    repo_validator_ssh = re.compile("^ssh://git@gitlab.cern.ch:7999/[a-zA-Z0-9_%-]+/[a-zA-Z0-9_%/-]+.git$")
-    repo_validator_https = re.compile("^https://gitlab.cern.ch/[a-zA-Z0-9_%-]+/[a-zA-Z0-9_%/-]+.git$")
-    repo_validator_kerb = re.compile("^https://:@gitlab.cern.ch:8443/[a-zA-Z0-9_%-]+/[a-zA-Z0-9_%/-]+.git$")
-
-    if repo_validator_https.match(repo_url) or \
-       repo_validator_ssh.match(repo_url) or \
-       repo_validator_kerb.match(repo_url):
-        # Repo URL explicit: should be created by the user
-        return repo_url, False
-
-    elif repo_url == '' or repo_url == "default":
-        # Default case: we need to resolve this
-        repo_url = "default"
-
-    else:
-        # Any other case is invalid
-        # NOTE: "no-gitlab" should be intercepted separately
-        return None, False
-
-    # Resolve 'default' by protocol
-    if repo_url == 'default':
-
-        if upload_protocol == 'ssh':
-            repo_url = "ssh://git@gitlab.cern.ch:7999/{}/{}.git".format(GROUP_NAME, project_name)
-        elif upload_protocol == 'https':
-            repo_url = "https://gitlab.cern.ch/{}/{}.git".format(GROUP_NAME, project_name)
-        elif upload_protocol == 'kerberos':
-            repo_url = "https://:@gitlab.cern.ch:8443/{}/{}.git".format(GROUP_NAME, project_name)
-        else:
-            raise ValueError("Upload protocol not recognized: '{}'".format(upload_protocol))
-
-    return repo_url, True
-
-
-def validate_gitlab(gitlab, repo, interactive, upload_protocol, clone_protocol, project_name) \
-        -> Mapping[str, Union[str, bool, None]]:
+def validate_gitlab(gitlab, repo_type, upload_protocol, clone_protocol, project_name, cern_id) -> Optional[str]:
     """
     Ensures all the parameters related to GitLab are in order.
-    :return: a dictionary containing the 'gitlab', 'create_repo' and 'repo_url' values
+    :return: a dictionary containing the 'gitlab' and 'repo_url' values
     """
     if gitlab:
-
         # Make sure the upload protocol is specified
         if upload_protocol is None:
             upload_protocol = clone_protocol
 
-        resolver = lambda r: resolve_gitlab_repo_url(r, upload_protocol, project_name)
-        question = "Press ENTER to create a GitLab repository for your project under 'bisw-python', or enter your " \
-                   "existing GitLab repository address here (or type 'no-gitlab' to keep your repository local):"
-        neg_feedback = "Invalid GitLab repository URL."
-        pos_feedback = "The project GitLab repository address is set to: \033[0;32m{}\033[0;m"
-
-        # Deal with CLI passing 'no-gitlab'
-        if repo == "no-gitlab":
-            cli.positive_feedback(pos_feedback.format(repo))
-            gitlab = False
-            repo = None
-            create_repo = False
-
+        if repo_type == 'operational':
+            group = "bisw-python"
+        elif repo_type == 'test':
+            group = cern_id
         else:
-            # This part is very similar to validate_as_arg_or_ask, but it returns the resolved values instead of a bool
-            # TODO make validate_as_arg_or_ask able to deal with this case too
-            if repo is not None or not interactive:
-                repo, create_repo = resolver(repo)
-                if repo is None:
-                    raise ValueError(neg_feedback)
-                cli.positive_feedback(pos_feedback.format(repo))
-            else:
-                while repo is None:
-                    repo = cli.ask_input(question).strip()
-                    if repo == "no-gitlab":
-                        break
-                    repo, create_repo = resolver(repo)
-                    if repo is None:
-                        cli.negative_feedback(neg_feedback)
-                cli.positive_feedback(pos_feedback.format(repo))
+            raise ValueError("Repository type not recognized: '{}'".format(repo_type))
 
-            if repo == "no-gitlab":
-                gitlab = False
-                repo = None
-                create_repo = False
-
+        if upload_protocol == 'ssh':
+            repo_url = "ssh://git@gitlab.cern.ch:7999/{}/{}.git".format(group, project_name)
+        elif upload_protocol == 'https':
+            repo_url = "https://gitlab.cern.ch/{}/{}.git".format(group, project_name)
+        elif upload_protocol == 'kerberos':
+            repo_url = "https://:@gitlab.cern.ch:8443/{}/{}.git".format(group, project_name)
+        else:
+            raise ValueError("Upload protocol not recognized: '{}'".format(upload_protocol))
     else:
-        create_repo = False
-        repo = None
+        repo_url = None
 
-    return {
-        "gitlab": gitlab,
-        "repo_url": repo,
-        "create_repo": create_repo,
-    }
+    return repo_url
 
 
-def validate_cern_id(cern_id: str) -> Optional[PhonebookEntry]:
+def validate_cern_id(cern_id: str) -> Tuple[Optional[PhonebookEntry], bool]:
     """
     Uses the Phonebook utilities to validate CERN IDs and retrieve their data.
     :param cern_id: The CERN ID of the user
@@ -252,6 +190,7 @@ def validate_cern_id(cern_id: str) -> Optional[PhonebookEntry]:
     """
     phonebook = Phonebook(cern_id)
     entries = phonebook.query_data()
+    print(entries)
     if len(entries) == 1 and cern_id in [login.cern_id for login in entries[0].login_list]:
-        return entries[0]
-    return None
+        return entries[0], True
+    return None, False
